Changed method and variable naming
Previously, my delete and mark operations used a variabled called
oneBased to indicate the taskNumber being marked. I was told that
is unintuitive and hence changed it to become taskNumber. For my
method to display the taskList, i named it render and similarly,
to make it sound more intuitive, i changed it to displayList.

This improves overall readability of code.